Move loading state backdrop to own component
